+++ server/src/main/java/org/opensearch/common/settings/IndexScopedSettings.java
@@ -185,6 +185,7 @@ public final class IndexScopedSettings extends AbstractScopedSettings {
                 IndexSettings.INDEX_SEARCH_IDLE_AFTER,
                 IndexSettings.INDEX_SEARCH_THROTTLED,
                 IndexSettings.INDEX_UNREFERENCED_FILE_CLEANUP,
+                IndexSettings.INDEX_QUERY_CACHE_MIN_FREQUENCY,
                 IndexFieldDataService.INDEX_FIELDDATA_CACHE_KEY,
                 FieldMapper.IGNORE_MALFORMED_SETTING,
                 FieldMapper.COERCE_SETTING,